requirements.txt
lxml==5.4.0
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-docx==1.1.2
shiboken6==6.9.0
typing_extensions==4.13.2
